+++ CHANGES.md
@@ -17,6 +17,9 @@ v0.6 (unreleased)
   ``get_default_factory`` functions are now deprecated since it is possible to
   achieve this solely with priorities. [#719]
 
+* Improved cube slider to include editable slice value as well as
+  first/previous/next/last buttons. [#690]
+
 * Registering data factories should now always be done with the
   ``@data_factory`` decorator, and not by adding functions to
   ``__factories__``, as was possible in early versions of Glue. [#724]